funcionalidades nuevas

+++ app/src/main/java/obligatorio/ort/obligatorio/estacionamiento/Calificacion.java
@@ -2,14 +2,14 @@ package obligatorio.ort.obligatorio.estacionamiento;
 
 public class Calificacion {
 
-    private int calificacion;
+    private double calificacion;
     private String comentario;
 
-    public int getCalificacion() {
+    public double getCalificacion() {
         return calificacion;
     }
 
-    public void setCalificacion(int calificacion) {
+    public void setCalificacion(double calificacion) {
         this.calificacion = calificacion;
     }
